int64 for file_length
File size is now using int64 for length (very big files)

darcs-hash:20061019150550-1b86b-678ba128753aa5b82aee85de33404fa3249f2da5

+++ http/http_frame.ml
@@ -35,7 +35,7 @@ module type HTTP_CONTENT =
     val content_of_string : string -> t Lwt.t
 
     (** convert a content type into a thread returning its size,etag,stream*)
-    val stream_of_content : t -> (int * string * stream) Lwt.t
+    val stream_of_content : t -> (int64 * string * stream) Lwt.t
   end
